+++ bin/github_download.dart
@@ -1,19 +1,21 @@
 
 import 'dart:io';
 
+const pactVersion = '0.1.1';
+
 String _getDownloadLink() {
   if (Platform.isLinux) {
-    return 'https://github.com/pact-foundation/pact-reference/releases/download/libpact_mock_server_ffi-v0.0.17/libpact_mock_server_ffi-linux-x86_64.so.gz';
+    return 'https://github.com/pact-foundation/pact-reference/releases/download/libpact_mock_server_ffi-v$pactVersion/libpact_mock_server_ffi-linux-x86_64.so.gz';
   } else if (Platform.isMacOS) {
-    return 'https://github.com/pact-foundation/pact-reference/releases/download/libpact_mock_server_ffi-v0.0.17/libpact_mock_server_ffi-osx-x86_64.dylib.gz';
+    return 'https://github.com/pact-foundation/pact-reference/releases/download/libpact_mock_server_ffi-v$pactVersion/libpact_mock_server_ffi-osx-x86_64.dylib.gz';
   } else if (Platform.isWindows) {
-    return 'https://github.com/pact-foundation/pact-reference/releases/download/libpact_mock_server_ffi-v0.0.17/libpact_mock_server_ffi-windows-x86_64.dll.gz';
+    return 'https://github.com/pact-foundation/pact-reference/releases/download/libpact_mock_server_ffi-v$pactVersion/libpact_mock_server_ffi-windows-x86_64.dll.gz';
   }
   throw Exception('Unsupported platform: ${Platform.operatingSystem}');
 }
 
-final _finalPath = Directory.systemTemp.path + '/libpact_mock_server_ffi-v0.0.17';
-final _downloadPath = Directory.systemTemp.path + '/libpact_mock_server_ffi-v0.0.17.gz';
+final _finalPath = '${Directory.systemTemp.path}/libpact_mock_server_ffi-v$pactVersion';
+final _downloadPath = '${Directory.systemTemp.path}/libpact_mock_server_ffi-v$pactVersion.gz';
 
 Future<String> downloadFromGithub() async {
   if (await File(_finalPath).exists()) {
@@ -25,31 +27,32 @@ Future<String> downloadFromGithub() async {
   return _finalPath;
 }
 
-void _unzip() async {
+Future<void> _unzip() async {
   print('Decompressing file $_downloadPath');
-  var bytes = await File(_downloadPath).readAsBytes();
-  var decoded = gzip.decode(bytes);
-  var writeStream = File(_finalPath).openWrite();
-  await writeStream.add(decoded);
+  final bytes = await File(_downloadPath).readAsBytes();
+  final decoded = gzip.decode(bytes);
+  print('Writing file $_finalPath');
+  final writeStream = File(_finalPath).openWrite();
+  writeStream.add(decoded);
   await writeStream.close();
 }
 
-void _download() async {
-  var file = File(_downloadPath);
+Future<void> _download() async {
+  final file = File(_downloadPath);
   if (await file.exists()) {
     print('GZ file exists');
     return;
   }
 
-  var link = _getDownloadLink();
+  final link = _getDownloadLink();
   print('Downloading gz file: $link');
-  var request = await HttpClient().getUrl(Uri.parse(link));
-  var response = await request.close();
-  var streamConsumer = file.openWrite();
+  final request = await HttpClient().getUrl(Uri.parse(link));
+  final response = await request.close();
+  final streamConsumer = file.openWrite();
   await response.pipe(streamConsumer);
   await streamConsumer.close();
 }
 
-void main() {
-  downloadFromGithub();
+Future<void> main() async {
+  await downloadFromGithub();
 }