+++ lib/src/pact_contract_dto.g.dart
@@ -6,200 +6,94 @@ part of 'pact_contract_dto.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-Pact _$PactFromJson(Map<String, dynamic> json) {
-  return Pact()
-    ..provider = json['provider'] == null
-        ? null
-        : Provider.fromJson(json['provider'] as Map<String, dynamic>)
-    ..consumer = json['consumer'] == null
-        ? null
-        : Consumer.fromJson(json['consumer'] as Map<String, dynamic>)
-    ..interactions = (json['interactions'] as List)
-        ?.map((e) =>
-            e == null ? null : Interaction.fromJson(e as Map<String, dynamic>))
-        ?.toList()
-    ..metadata = json['metadata'] == null
-        ? null
-        : Metadata.fromJson(json['metadata'] as Map<String, dynamic>);
-}
+Pact _$PactFromJson(Map<String, dynamic> json) => Pact()
+  ..provider = Provider.fromJson(json['provider'] as Map<String, dynamic>)
+  ..consumer = Consumer.fromJson(json['consumer'] as Map<String, dynamic>)
+  ..interactions = (json['interactions'] as List<dynamic>)
+      .map((e) => Interaction.fromJson(e as Map<String, dynamic>))
+      .toList()
+  ..metadata = Metadata.fromJson(json['metadata'] as Map<String, dynamic>);
 
-Map<String, dynamic> _$PactToJson(Pact instance) {
-  final val = <String, dynamic>{};
+Map<String, dynamic> _$PactToJson(Pact instance) => <String, dynamic>{
+      'provider': instance.provider,
+      'consumer': instance.consumer,
+      'interactions': instance.interactions,
+      'metadata': instance.metadata,
+    };
 
-  void writeNotNull(String key, dynamic value) {
-    if (value != null) {
-      val[key] = value;
-    }
-  }
+Metadata _$MetadataFromJson(Map<String, dynamic> json) => Metadata()
+  ..pactSpecification =
+      Map<String, String>.from(json['pactSpecification'] as Map)
+  ..pactDart = Map<String, String>.from(json['pact-dart'] as Map);
 
-  writeNotNull('provider', instance.provider);
-  writeNotNull('consumer', instance.consumer);
-  writeNotNull('interactions', instance.interactions);
-  writeNotNull('metadata', instance.metadata);
-  return val;
-}
+Map<String, dynamic> _$MetadataToJson(Metadata instance) => <String, dynamic>{
+      'pactSpecification': instance.pactSpecification,
+      'pact-dart': instance.pactDart,
+    };
 
-Metadata _$MetadataFromJson(Map<String, dynamic> json) {
-  return Metadata()
-    ..pactSpecification =
-        (json['pactSpecification'] as Map<String, dynamic>)?.map(
-      (k, e) => MapEntry(k, e as String),
-    )
-    ..pactDart = (json['pact-dart'] as Map<String, dynamic>)?.map(
-      (k, e) => MapEntry(k, e as String),
-    );
-}
+Interaction _$InteractionFromJson(Map<String, dynamic> json) => Interaction()
+  ..description = json['description'] as String
+  ..request = Request.fromJson(json['request'] as Map<String, dynamic>)
+  ..response = Response.fromJson(json['response'] as Map<String, dynamic>)
+  ..providerStates = (json['providerStates'] as List<dynamic>)
+      .map((e) => ProviderState.fromJson(e as Map<String, dynamic>))
+      .toList();
 
-Map<String, dynamic> _$MetadataToJson(Metadata instance) {
-  final val = <String, dynamic>{};
+Map<String, dynamic> _$InteractionToJson(Interaction instance) =>
+    <String, dynamic>{
+      'description': instance.description,
+      'request': instance.request,
+      'response': instance.response,
+      'providerStates': instance.providerStates,
+    };
 
-  void writeNotNull(String key, dynamic value) {
-    if (value != null) {
-      val[key] = value;
-    }
-  }
+ProviderState _$ProviderStateFromJson(Map<String, dynamic> json) =>
+    ProviderState()
+      ..name = json['name'] as String
+      ..params = json['params'] as Map<String, dynamic>;
 
-  writeNotNull('pactSpecification', instance.pactSpecification);
-  writeNotNull('pact-dart', instance.pactDart);
-  return val;
-}
+Map<String, dynamic> _$ProviderStateToJson(ProviderState instance) =>
+    <String, dynamic>{
+      'name': instance.name,
+      'params': instance.params,
+    };
 
-Interaction _$InteractionFromJson(Map<String, dynamic> json) {
-  return Interaction()
-    ..description = json['description'] as String
-    ..request = json['request'] == null
-        ? null
-        : Request.fromJson(json['request'] as Map<String, dynamic>)
-    ..response = json['response'] == null
-        ? null
-        : Response.fromJson(json['response'] as Map<String, dynamic>)
-    ..providerStates = (json['providerStates'] as List)
-        ?.map((e) => e == null
-            ? null
-            : ProviderState.fromJson(e as Map<String, dynamic>))
-        ?.toList();
-}
+Response _$ResponseFromJson(Map<String, dynamic> json) => Response()
+  ..status = json['status'] as int
+  ..headers = Map<String, String>.from(json['headers'] as Map)
+  ..body = Body.fromJsonToBody(json['body']);
 
-Map<String, dynamic> _$InteractionToJson(Interaction instance) {
-  final val = <String, dynamic>{};
+Map<String, dynamic> _$ResponseToJson(Response instance) => <String, dynamic>{
+      'status': instance.status,
+      'headers': instance.headers,
+      'body': instance.body,
+    };
 
-  void writeNotNull(String key, dynamic value) {
-    if (value != null) {
-      val[key] = value;
-    }
-  }
+Request _$RequestFromJson(Map<String, dynamic> json) => Request()
+  ..method = json['method'] as String
+  ..path = json['path'] as String
+  ..query = Map<String, String>.from(json['query'] as Map)
+  ..headers = Map<String, String>.from(json['headers'] as Map)
+  ..body = Body.fromJsonToBody(json['body']);
 
-  writeNotNull('description', instance.description);
-  writeNotNull('request', instance.request);
-  writeNotNull('response', instance.response);
-  writeNotNull('providerStates', instance.providerStates);
-  return val;
-}
+Map<String, dynamic> _$RequestToJson(Request instance) => <String, dynamic>{
+      'method': instance.method,
+      'path': instance.path,
+      'query': instance.query,
+      'headers': instance.headers,
+      'body': instance.body,
+    };
 
-ProviderState _$ProviderStateFromJson(Map<String, dynamic> json) {
-  return ProviderState()
-    ..name = json['name'] as String
-    ..params = json['params'] as Map<String, dynamic>;
-}
+Provider _$ProviderFromJson(Map<String, dynamic> json) =>
+    Provider()..name = json['name'] as String;
 
-Map<String, dynamic> _$ProviderStateToJson(ProviderState instance) {
-  final val = <String, dynamic>{};
+Map<String, dynamic> _$ProviderToJson(Provider instance) => <String, dynamic>{
+      'name': instance.name,
+    };
 
-  void writeNotNull(String key, dynamic value) {
-    if (value != null) {
-      val[key] = value;
-    }
-  }
+Consumer _$ConsumerFromJson(Map<String, dynamic> json) =>
+    Consumer()..name = json['name'] as String;
 
-  writeNotNull('name', instance.name);
-  writeNotNull('params', instance.params);
-  return val;
-}
-
-Response _$ResponseFromJson(Map<String, dynamic> json) {
-  return Response()
-    ..status = json['status'] as int
-    ..headers = (json['headers'] as Map<String, dynamic>)?.map(
-      (k, e) => MapEntry(k, e as String),
-    )
-    ..body = Body.fromJsonToBody(json['body']);
-}
-
-Map<String, dynamic> _$ResponseToJson(Response instance) {
-  final val = <String, dynamic>{};
-
-  void writeNotNull(String key, dynamic value) {
-    if (value != null) {
-      val[key] = value;
-    }
-  }
-
-  writeNotNull('status', instance.status);
-  writeNotNull('headers', instance.headers);
-  writeNotNull('body', instance.body);
-  return val;
-}
-
-Request _$RequestFromJson(Map<String, dynamic> json) {
-  return Request()
-    ..method = json['method'] as String
-    ..path = json['path'] as String
-    ..query = (json['query'] as Map<String, dynamic>)?.map(
-      (k, e) => MapEntry(k, e as String),
-    )
-    ..headers = (json['headers'] as Map<String, dynamic>)?.map(
-      (k, e) => MapEntry(k, e as String),
-    )
-    ..body = Body.fromJsonToBody(json['body']);
-}
-
-Map<String, dynamic> _$RequestToJson(Request instance) {
-  final val = <String, dynamic>{};
-
-  void writeNotNull(String key, dynamic value) {
-    if (value != null) {
-      val[key] = value;
-    }
-  }
-
-  writeNotNull('method', instance.method);
-  writeNotNull('path', instance.path);
-  writeNotNull('query', instance.query);
-  writeNotNull('headers', instance.headers);
-  writeNotNull('body', instance.body);
-  return val;
-}
-
-Provider _$ProviderFromJson(Map<String, dynamic> json) {
-  return Provider()..name = json['name'] as String;
-}
-
-Map<String, dynamic> _$ProviderToJson(Provider instance) {
-  final val = <String, dynamic>{};
-
-  void writeNotNull(String key, dynamic value) {
-    if (value != null) {
-      val[key] = value;
-    }
-  }
-
-  writeNotNull('name', instance.name);
-  return val;
-}
-
-Consumer _$ConsumerFromJson(Map<String, dynamic> json) {
-  return Consumer()..name = json['name'] as String;
-}
-
-Map<String, dynamic> _$ConsumerToJson(Consumer instance) {
-  final val = <String, dynamic>{};
-
-  void writeNotNull(String key, dynamic value) {
-    if (value != null) {
-      val[key] = value;
-    }
-  }
-
-  writeNotNull('name', instance.name);
-  return val;
-}
+Map<String, dynamic> _$ConsumerToJson(Consumer instance) => <String, dynamic>{
+      'name': instance.name,
+    };